% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_12S.R
\name{assignment_12S}
\alias{assignment_12S}
\title{12S Assignment}
\usage{
assignment_12S(
  qiime_asvtab = qiime.asvtab.12S,
  ref_fasta = ref.12S,
  out_dir = qiime.dir.12S,
  tryRC = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{qiime_asvtab}{The qiime-formatted ASV table}

\item{ref_fasta}{The path to the 12S reference FASTA}

\item{out_dir}{The path to the output directory}

\item{tryRC}{TRUE or FALSE; whether to check reverse complements}

\item{verbose}{TRUE or FALSE; whether to output messages}
}
\value{
A taxonomy table, formatted as a character matrix; a list of statistics regarding assignment,
and a CSV of NAs saved to the output directory.
}
\description{
This function creates a taxonomy table for 12S ASVs.
}
