% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_trnL.R
\name{assignment_trnL}
\alias{assignment_trnL}
\title{trnL Assignment}
\usage{
assignment_trnL(
  qiime_asvtab = qiime.asvtab.trnL,
  ref_fasta = ref.trnL,
  out_dir = qiime.dir.trnL,
  tryRC = TRUE,
  manual_taxids = NULL,
  verbose = TRUE
)
}
\arguments{
\item{qiime_asvtab}{The qiime-formatted ASV table}

\item{ref_fasta}{The path to the trnL reference FASTA}

\item{out_dir}{The path to the output directory}

\item{tryRC}{TRUE or FALSE; whether to check reverse complements}

\item{manual_taxids}{NULL; manual additions to be made}

\item{verbose}{TRUE or FALSE; whether to output messages}
}
\value{
A taxonomy table, formatted as a character matrix; a list of statistics regarding assignment,
and a CSV of NAs saved to the output directory.
}
\description{
This function creates a taxonomy table for trnL ASVs.
}
