% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_phyloseq.R
\name{filter_phyloseq}
\alias{filter_phyloseq}
\title{Filter Phyloseq by Sample Metadata Field}
\usage{
filter_phyloseq(ps, field, value)
}
\arguments{
\item{ps}{Phyloseq object.}

\item{field}{Character; name of the column in sample_data to filter on.}

\item{value}{The value to match; samples where \code{field == value} are
retained.}
}
\value{
Phyloseq object containing only the matching samples.
}
\description{
Subsets a phyloseq object to samples where a specified
sample_data field equals a given value. A convenience wrapper around
\code{phyloseq::prune_samples}.
}
