% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_table_seqs.R
\name{join_table_seqs}
\alias{join_table_seqs}
\title{Join Table Sequences}
\usage{
join_table_seqs(feature_table, sequence_hash)
}
\arguments{
\item{feature_table}{A QIIME2-formatted feature table}

\item{sequence_hash}{A QIIME2-formatted sequence hash table}
}
\value{
A feature table with a sequence hash table joined to it
}
\description{
This function joins a sequence hash table with a feature
table from QIIME2.
}
