% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plot.R
\name{pca_plot}
\alias{pca_plot}
\title{PCA plot and biplot}
\usage{
pca_plot(
  ps,
  colorVar = NULL,
  colorName = NULL,
  nTaxa = 10,
  customColors = NULL,
  customGradient = NULL,
  mid = "mean",
  xPC = 1,
  yPC = 2,
  ellipse = FALSE,
  bplab = NULL
)
}
\arguments{
\item{ps}{clr transformed and filtered data}

\item{colorVar}{variable from samdf to color samples by}

\item{colorName}{what to display variable name as in legend}

\item{nTaxa}{number of taxa to display}

\item{customColors}{optional named vector of colors}

\item{customGradient}{optional df of low/high colors for gradient}

\item{mid}{for customGradient -- midpoint to median, mean, middle of range}

\item{xPC}{Principal Component for x-axis}

\item{yPC}{Principal Component for y-axis}

\item{ellipse}{optional add centroid ellipses}

\item{bplab}{name of column that you want to use for labeling the biplot}
}
\value{
A named list with six elements: \code{pca.df} (samdf with all PCs
added as columns), \code{pca.biplot} (PCA biplot with \code{nTaxa}
factor loadings displayed), \code{loadings} (data frame of loadings in
PCA space), \code{pca.output} (raw output of \code{prcomp()}),
\code{scree.table} (table of scree-plot numbers), and \code{scree.plot}
(the scree plot).
}
\description{
This function runs a Principal Component Analysis.
}
