% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_asv_length_hist.R
\name{plot_asv_length_hist}
\alias{plot_asv_length_hist}
\title{Plot ASV Length Histograms}
\usage{
plot_asv_length_hist(
  marker,
  asvtab = get(paste0("qiime.asvtab.", marker), envir = parent.frame()),
  out_dir = get(paste0("qiime.dir.", marker), envir = parent.frame()),
  proj = project,
  bin_size = 2
)
}
\arguments{
\item{marker}{The marker gene, either "trnL" or "12S"}

\item{asvtab}{The ASV table (samples x ASVs). Defaults to looking up
\code{qiime.asvtab.<marker>} in the calling environment.}

\item{out_dir}{Path to the output directory for saving the plot. Defaults to
looking up \code{qiime.dir.<marker>} in the calling environment.}

\item{proj}{The name of the project}

\item{bin_size}{The bin size of the histogram}
}
\value{
A histogram of ASV lengths and a dataframe of lengths
}
\description{
This function plots a histogram of ASV lengths.
}
