% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_taxonomy.R
\name{update_taxonomy}
\alias{update_taxonomy}
\title{Update Taxonomy}
\usage{
update_taxonomy(ps, reference, marker)
}
\arguments{
\item{ps}{A phyloseq object}

\item{reference}{A file path to a FASTA-formatted reference}

\item{marker}{The marker name as a string}
}
\value{
Updated phyloseq object
}
\description{
This function updates the taxonomic assignments of a
phyloseq object.
}
