% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_asv_length_hist.R
\name{plot_asv_length_hist}
\alias{plot_asv_length_hist}
\title{Plot ASV Length Histograms}
\usage{
plot_asv_length_hist(marker, proj = project, bin_size = 2)
}
\arguments{
\item{marker}{The marker gene, either "trnL" or "12S"}

\item{proj}{The name of the project}

\item{bin_size}{The bin size of the histogram}
}
\value{
A histogram of ASV lengths and a dataframe of lengths
}
\description{
This function plots a histogram of ASV lengths.
}
