% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_qiime_run.R
\name{process_qiime_run}
\alias{process_qiime_run}
\title{Process QIIME Run}
\usage{
process_qiime_run(marker, proj = project, download_mode = mode)
}
\arguments{
\item{marker}{The marker gene, either "trnL" or "12S"}

\item{proj}{The name of the project}

\item{dl_mode}{If all files from the QIIME run were downloaded or just the necessary ones, either "full" or "minimal"}
}
\value{
A track table and a plot of read counts, optionally
}
\description{
This function optionally unzips files, creates a track table, and outputs a plot of read counts depending on inputs.
}
