% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_controls.R
\name{qc_controls}
\alias{qc_controls}
\title{Quality Controls}
\usage{
qc_controls(
  marker,
  proj = project,
  ps = NULL,
  control_species = NULL,
  out_dir = NULL
)
}
\arguments{
\item{marker}{The marker gene, either "trnL" or "12S"}

\item{proj}{The name of the project}

\item{ps}{The phyloseq object}

\item{control_species}{A character vector of positive control species names}

\item{out_dir}{The path to the output directory}
}
\value{
A plot of control compositions, a plot of samples with positive controls detected, and a plate map
}
\description{
This function outputs quality control plots as relate to controls and possible contamination.
}
