% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate_to_folder.R
\name{truncate_to_folder}
\alias{truncate_to_folder}
\title{Truncate to Folder}
\usage{
truncate_to_folder(path, folder)
}
\arguments{
\item{path}{A file path}

\item{folder}{The folder of interest}
}
\value{
A file path truncated to the folder of interest
}
\description{
This function allows you to truncate a given path to a folder of interest.
}
